+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [artifact-manager] IAC-733 / Add an Option for Overwriting Existing vRLI Content Packs.
 * [artifact-manager] IAC-741 / mvn vrealize:clean now will not fail if not supported
 * [artifact-manager] IAC-745 / VRLI Alerts Fallback Object Set to LogInsight Only During Push
+* [artifact-manager] IAC-693 / vRBT to support new content sharing policies.
 
 ## v2.29.2 - 24 Feb 2023